Norm'd

+++ ft_strncmp.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-int		ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	size_t	i;
 
@@ -8,8 +8,8 @@ int		ft_strncmp(const char *s1, const char *s2, size_t n)
 	while (s1[i] && i < n && s2[i])
 	{
 		if (s1[i] != s2[i])
-			return (s1[i] - s2[i]);
+			return ((int)(s1[i] - s2[i]));
 		i++;
 	}
-	return (int)(s1[i] - s2[i]);
+	return ((int)(s1[i] - s2[i]));
 }
